+++ src/internal/MicrogrammarSpecParser.ts
@@ -3,25 +3,22 @@ import { Concat, toMatchingLogic } from "../matchers/Concat";
 import { Literal } from "../Primitives";
 
 import { WhiteSpaceHandler } from "../Config";
+import { FromStringOptions } from "../FromStringOptions";
 import { isPatternMatch } from "../PatternMatch";
 import { Break } from "./Break";
+import { CompleteFromStringOptions, completeWithDefaults } from "./CompleteFromStringOptions";
 import { exactMatch } from "./ExactMatch";
 import { MicrogrammarSpec, specGrammar } from "./SpecGrammar";
 
 /**
- * Use this token when you want an anonymous, unbound stream of content to skip
- * @type {string}
- */
-const DiscardToken = "⤞";
-
-/**
  * Convenient function to create a microgrammar from a spec within another grammar
  * @param spec string spec
- * @param elements
+ * @param components
+ * @param options
  * @returns {Concat}
  */
-export function sentence(spec: string, elements: object = {}): Concat {
-    return new MicrogrammarSpecParser().fromString(spec, elements);
+export function fromString(spec: string, components: object = {}, options: FromStringOptions = {}): Concat {
+    return new MicrogrammarSpecParser().fromString(spec, components, options);
 }
 
 /**
@@ -31,22 +28,23 @@ export class MicrogrammarSpecParser {
 
     private anonFieldCount = 0;
 
-    public fromString(spec: string, elements: object = {}): Concat {
-        spec = this.preprocess(spec);
-        const match = exactMatch<MicrogrammarSpec>(specGrammar, spec);
+    public fromString(spec: string, components: object, options: FromStringOptions): Concat {
+        const optionsToUse = completeWithDefaults(options);
+        spec = this.preprocess(spec, optionsToUse);
+        const match = exactMatch<MicrogrammarSpec>(specGrammar(optionsToUse), spec);
         if (!isPatternMatch(match)) {
             throw new Error(`Unable to parse microgrammar: ${spec}`);
         }
         const matcherSequence1 = this.definitionSpecsFromMicrogrammarSpec(match,
-            (elements as WhiteSpaceHandler).$consumeWhiteSpaceBetweenTokens !== false);
-        const matcherSequence2 = this.populateSpecifiedElements(elements, matcherSequence1);
+            (components as WhiteSpaceHandler).$consumeWhiteSpaceBetweenTokens !== false);
+        const matcherSequence2 = this.populateSpecifiedElements(components, matcherSequence1);
         const matcherSequence3 = this.inferUnspecifiedElements(matcherSequence2);
         const definitions = this.definitionsFromSpecs(spec, matcherSequence3);
         const concat = new Concat(definitions);
         // Copy config to Concat
-        for (const key in elements) {
+        for (const key in components) {
             if (key.charAt(0) === "$") {
-                concat[key] = elements[key];
+                concat[key] = components[key];
             }
         }
         return concat;
@@ -56,14 +54,15 @@ export class MicrogrammarSpecParser {
      * Given a spec, replace all the DiscardToken instances with a named,
      * but unbound, matcher spec
      * @param spec spec to preprocess before parsing
+     * @param optionsToUse options
      * @returns {string}
      */
-    private preprocess(spec: string): string {
-        const split = spec.split(DiscardToken);
+    private preprocess(spec: string, optionsToUse: CompleteFromStringOptions): string {
+        const split = spec.split(optionsToUse.ellipsis);
         let joined = "";
         for (let i = 0; i < split.length; i++) {
             if (i > 0) {
-                joined += "${_discard" + i + "}";
+                joined += `${optionsToUse.componentPrefix}{_discard${i}}`;
             }
             joined += split[i];
         }