+++ test/fromString/SpecGrammarTest.ts
@@ -1,4 +1,5 @@
 import assert = require("power-assert");
+import { DefaultFromStringOptions } from "../../src/internal/CompleteFromStringOptions";
 import { exactMatch } from "../../src/internal/ExactMatch";
 import { MicrogrammarSpec, specGrammar } from "../../src/internal/SpecGrammar";
 import {isPatternMatch} from "../../src/PatternMatch";
@@ -7,7 +8,7 @@ describe("SpecGrammar", () => {
 
     it("can parse a series of literals and references", () => {
         const microgrammarSpecString = "->${fruit}<-";
-        const specMatch = exactMatch<MicrogrammarSpec>(specGrammar, microgrammarSpecString);
+        const specMatch = exactMatch<MicrogrammarSpec>(specGrammar(DefaultFromStringOptions), microgrammarSpecString);
         if (isPatternMatch(specMatch)) {
             assert(specMatch.these.length === 1);
             assert.deepEqual(specMatch.matchedStructure(),
@@ -24,7 +25,7 @@ describe("SpecGrammar", () => {
 
     it("can parse three references in succession", () => {
         const specString = "->${fruit}${arrow}${drink}!";
-        const specMatch = exactMatch<MicrogrammarSpec>(specGrammar, specString);
+        const specMatch = exactMatch<MicrogrammarSpec>(specGrammar(DefaultFromStringOptions), specString);
         if (isPatternMatch(specMatch)) {
             assert(specMatch.these.length === 3);
             assert(specMatch.these[0].element.elementName === "fruit");