+++ CHANGELOG.md
@@ -11,9 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
--   Added nesting support via `sentence` function for `Microgrammar.fromString`
--   Allow `⤞` character in `Microgrammar.fromString` specs to skip over intervening content
-    without binding
+-   Added nesting support via `fromString` function for `Microgrammar.fromString`
+-   Introduced `FromStringOptions` to customize `Microgrammar.fromString` behavior
+-   Added `FromStringOptions.ellipsis` option for `Microgrammar.fromString` specs to skip over intervening content
+    without binding. Defaults to `...`.
+-   Added `FromStringOptions.componentPrefix` option for `Microgrammar.fromString` specs to 
+    specify component prefix. Defaults to former `$`.
 
 ### Changed